modify the way project files are exported so that may be imported into other projects more smoothly
