+++ index.ts
@@ -1,0 +1,9 @@
+/**
+ * @module
+ * @description
+ * Entry point for all public APIs of the appworks package.
+ */
+
+export * from './src/index';
+
+// This file only reexports content of the `src` folder. Keep it that way.
