+++ src/index.ts
@@ -1,0 +1,39 @@
+/**
+ * @module
+ * @description
+ * Entry point from which you should import all public appworks APIs.
+ */
+
+export {AWPlugin} from './plugin';
+export {AWProxy, __aw_plugin_proxy} from './proxy';
+export {Util} from './util';
+export {AWAccelerometer} from './plugins/accelerometer';
+export {AWAppManager} from './plugins/app-manager';
+export {AWAuth} from './plugins/auth';
+export {AWCache, PersistentStorage} from './plugins/cache';
+export {AWCamera} from './plugins/camera';
+export {AWCompass} from './plugins/compass';
+export {AWComponent} from './plugins/component';
+export {AWContacts} from './plugins/contacts';
+export {AsyncStorage, DesktopStorage} from './plugins/desktop-storage';
+export {AWDevice} from './plugins/device';
+export {AWFileChooser} from './plugins/file-chooser';
+export {AWFileSystem, FileDetails, SaveDialogOptions, FileDialogOptions, DesktopHostFileSystem} from './plugins/file-system';
+export {AWFileTransfer} from './plugins/file-transfer';
+export {AWFinder} from './plugins/finder';
+export {AWHeader, AWHeaderBar} from './plugins/header';
+export {AWKeyboard} from './plugins/keyboard';
+export {LocalFileSystem} from './plugins/local-file-system';
+export {AWLocation} from './plugins/location';
+export {AWMedia} from './plugins/media';
+export {AWMediaCapture} from './plugins/media-capture';
+export {AWMenu} from './plugins/menu';
+export {AWNotificationManager} from './plugins/notifications';
+export {AWOfflineManager} from './plugins/offline-manager';
+export {OnDeviceStorage} from './plugins/on-device-storage';
+export {AWPage} from './plugins/page';
+export {AWQRReader} from './plugins/qr-reader';
+export {SecureStorage, AWSecureStorage} from './plugins/secure-storage';
+export {AWStorage} from './plugins/storage';
+export {AWVibration} from './plugins/vibration';
+export {AWWebView} from './plugins/webview';
