+++ tsconfig.json
@@ -11,10 +11,5 @@
   "include": [
     "./src/**/*.ts",
     "./src/**/*.tsx"
-  ],
-  "exclude": [
-    "node_modules",
-    "src/**/tests/*.ts",
-    "src/**/tests/*.tsx"
   ]
 }
